Allow overriding oAuth scopes for NAP pools
The defaults I included come from the scopes I observed in a cluster I
stood up when no scopes were specified. I am assuming these are GKE
defaults. This does not match the default scopes for normal node pools
in this Terraform module, so it may not be the correct choice.